add back to inv all

+++ client/src/components/AccessoryCart.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { useNavigate } from "react-router-dom";
 
 function AccessoryCart({
   cartAccessoryObj,
@@ -6,23 +7,66 @@ function AccessoryCart({
   setAccessoryCart,
   setErrors,
 }) {
+  const navigate = useNavigate();
+
   const handleDeleteAccessory = (id) => {
     const headers = {
       "Content-Type": "application/json",
       Authorization: `Bearer ${localStorage.token}`,
     };
 
-    fetch(`http://localhost:3000/user_accessories/${id}`, {
-      method: "DELETE",
-      headers: headers,
-    }).then((r) => {
-      if (r.ok) {
-        console.log("removed from cart");
-        setAccessoryCart(accessoryCart.filter((a) => a.id !== id));
-      } else {
-        r.json().then((err) => setErrors(err.errors));
-      }
-    });
+    if (!localStorage.getItem("currentUserId")) {
+      navigate("/login");
+    } else {
+      fetch(
+        `http://localhost:3000/accessories/${cartAccessoryObj.accessory_id}`,
+        {
+          method: "GET",
+          headers: headers,
+        }
+      ).then((r) => {
+        if (r.ok) {
+          r.json();
+        } else {
+          r.json().then((error) => setErrors(error.errors));
+          navigate("/login");
+        }
+      });
+
+      console.log(cartAccessoryObj.add_back);
+
+      const addBackToInvBody = {
+        count: cartAccessoryObj.add_back,
+      };
+      console.log(addBackToInvBody);
+
+      fetch(
+        `http://localhost:3000/accessories/${cartAccessoryObj.accessory_id}`,
+        {
+          method: "PATCH",
+          headers: headers,
+          body: JSON.stringify(addBackToInvBody),
+        }
+      ).then((r) => {
+        if (r.ok) {
+          console.log("added back to inventory");
+        } else {
+          r.json().then((err) => setErrors(err.errors));
+        }
+      });
+
+      fetch(`http://localhost:3000/user_accessories/${id}`, {
+        method: "DELETE",
+        headers: headers,
+      }).then((r) => {
+        if (r.ok) {
+          console.log("removed from cart");
+          setAccessoryCart(accessoryCart.filter((a) => a.id !== id));
+        } else {
+          r.json().then((err) => setErrors(err.errors));
+        }
+      });
+    }
   };
 
   return (
