+++ client/src/components/ApparelCart.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { useNavigate } from "react-router-dom";
 
 function ApparelCart({
   cartApparelObj,
@@ -6,23 +7,61 @@ function ApparelCart({
   setApparelCart,
   setErrors,
 }) {
+  const navigate = useNavigate();
+
   const handleDeleteApparel = (id) => {
     const headers = {
       "Content-Type": "application/json",
       Authorization: `Bearer ${localStorage.token}`,
     };
 
-    fetch(`http://localhost:3000/user_apparels/${id}`, {
-      method: "DELETE",
-      headers: headers,
-    }).then((r) => {
-      if (r.ok) {
-        console.log("removed from cart");
-        setApparelCart(apparelCart.filter((a) => a.id !== id));
-      } else {
-        r.json().then((err) => setErrors(err.errors));
-      }
-    });
+    if (!localStorage.getItem("currentUserId")) {
+      navigate("/login");
+    } else {
+      // console.log(cartSkiObj);
+      fetch(`http://localhost:3000/apparels/${cartApparelObj.apparel_id}`, {
+        method: "GET",
+        headers: headers,
+      }).then((r) => {
+        if (r.ok) {
+          r.json();
+        } else {
+          r.json().then((error) => setErrors(error.errors));
+          navigate("/login");
+        }
+      });
+
+      console.log(cartApparelObj.add_back);
+
+      const addBackToInvBody = {
+        count: cartApparelObj.add_back,
+      };
+      console.log(addBackToInvBody);
+
+      fetch(`http://localhost:3000/apparels/${cartApparelObj.apparel_id}`, {
+        method: "PATCH",
+        headers: headers,
+        body: JSON.stringify(addBackToInvBody),
+      }).then((r) => {
+        if (r.ok) {
+          console.log("added back to inventory");
+        } else {
+          r.json().then((err) => setErrors(err.errors));
+        }
+      });
+
+      fetch(`http://localhost:3000/user_apparels/${id}`, {
+        method: "DELETE",
+        headers: headers,
+      }).then((r) => {
+        if (r.ok) {
+          console.log("removed from cart");
+          setApparelCart(apparelCart.filter((a) => a.id !== id));
+        } else {
+          r.json().then((err) => setErrors(err.errors));
+        }
+      });
+    }
   };
 
   return (
